[block-in-inline] Change |NGInlineItem::IsEmptyItem()| to false
This patch changes |NGInlineItem::IsEmptyItem()| to false for
block-in-inline items.

For block-in-inline items, the accurate value cannot be
determined until it is laid out. The current usages are:
1. To determine if an inline box forces line box creation or
   not. This is only for open/close tag items.
2. |NGInlineLayoutAlgorithm::PositionLeadingFloats()| to
   find leading floats before starting laying out the inline
   formatting context.

For floats within a block-in-inline, e.g.:
```html
  <span>
    <div>aaa</div>
    <div style="float: left"></div>
```
This change prevents |PositionLeadingFloats()| to position the
float before |NGLineBreaker| handles the `<div>`.

Fixes DCHECK failures in |NGLineBreaker| for:
  external/wpt/css/CSS2/normal-flow/block-in-inline-float-between-001.xht
and 10 other tests.

Bug: 716930
Change-Id: I712c5f87d64febaa032e5a409704f4a9b1a08284
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3140800
Auto-Submit: Koji Ishii <kojii@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#918390}
NOKEYCHECK=True
GitOrigin-RevId: 67ccd0068a8f821a5841f92b4acf8dc1716e2204