Optimize data dependency for blink_web_tests
Some baselines are only useful for certain platforms. Explicitly list
such dependency so that we don't need download unnecessary baselines to
swarming bots. Added a presubmit check to ensure we do not miss any
folders under web_tests.

This should be able to reduce the overhead on the swarming bots. After
this change, we saved 38524 files or directories on Windows.
This is 20.8% of the web_tests folder.

Bug: 1230939
Change-Id: I805f03f9f002a05b0fb1456271aa9b383c9bd3c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3139556
Commit-Queue: Weizhong Xia <weizhong@google.com>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/main@{#918838}
NOKEYCHECK=True
GitOrigin-RevId: 8f26bc7ead0931b57048c34fe10e232a3db72e2c