[Web Payment] Enable SPC
Before this patch, SPC was disabled on trunk through a
"default-disabled" browser flag that controlled the renderer flag as
well.

This patch enables SPC browser flag for Mac and Windows on trunk. The
WPT expectations are updated on Mac and Windows because the
PaymentCredential Mojo service is not implemented in the content shell
and does not have test driver automation at this time, which makes the
error on Mac and Windows different from other platforms.

After this patch, SPC is enabled on trunk (currently M95) for Mac and
Windows through a browser flag that also controls the renderer flag as
well.

Intent to ship:
https://groups.google.com/a/chromium.org/g/blink-dev/c/U5K69fbA6SU/m/6Ipi3Q4zAgAJ

Bug: 1241540, 1124927
Change-Id: I967345eb4c041fcbf3d0b49012afd19c8e00dc19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3108106
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#918936}
NOKEYCHECK=True
GitOrigin-RevId: b964b814760d38ecf4199a15791e9fa1ed299353