Make web bundle and the inner subresource requests skip Service Worker
The explainer of "Subresource loading with Web Bundles" [*] doesn't
address the expected behavior of the service worker fetch interception.
And in the current implementation, if the service worker of the page
intercepts the web bundle request, the subsequent subresource requests
can’t be served from the web bundle.

As a short term solution, this CL sets the SkipServiceWorker flag of
the both requests. So the inner subresource can be served from the web
bundle.

As a long term solution, we need to figure out the ideal design of the
service worker integration. And we need to introduce an option argument
{skip_service_worker: false (default is true)} in future.

But as of now, we don’t have any good use case of the service worker
integration. So I think it is OK to land this short term solution.

[*] https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md

Bug: 1240424, https://github.com/WICG/webpackage/issues/676
Change-Id: I230daf58da72ecfa83b1756616a8a0697144acac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3145076
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Hayato Ito <hayato@chromium.org>
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919201}
NOKEYCHECK=True
GitOrigin-RevId: 0c6d4f9607924d0e6891627e72c73173fb679615