BackForwardCache: Disable when MediaSession's service is used
After this change, back-forward cache is blocked when MediaSession's
service exists, i.e., a MediaSession action handler is registered and a
audio or media plays. It is risky to put a page with a MediaSession
service into the back-forward cache, since the page's state can affect
a media control UI. This change adds this blocking behind a flag
(supported_features) for future experiments.

This change adds feature flags. We think that changing the playback
state without a MediaSession service should not be harmful for back-
forward cache. By adding the flags, we will be able to do experiments.

After this change, we have two new features:

 * BackForwardCacheMediaSessionPlaybackStateChange
 * BackForwardCacheMediaSessionService

These are independent and can take these pairs:

  * (false, false): MediaSession's playback state change doesn't allow
                    BFcache (default)
  * (true, false):  MediaSession's playback state change w/o a
                    MediaSession service allows BFcache
  * (true, true):   MediaSession's playback state change w and w/o a
                    MediaSession service allows BFcache

In theory there is a pair (false, true) but this doesn't make sense, as
a MediaSession service does nothing if the playback state is not
changed.

Bug: 1243599
Change-Id: Ieeb821a632ee5ccad345a802c4bd21fced641b2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3121727
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919212}
NOKEYCHECK=True
GitOrigin-RevId: 15254fa02faf5cd752a332912d76d5701d4f7b02