Ensure that EyeDropper is not created when the feature flag is disabled.
Before this CL we were relying on define guards to ensure that
the eye dropper is not created on not supported platforms[1].

This approach created an issue on ChromeOS, where we were attempting
to create an EyeDropper without having the feature flag enabled.

This change adds feature flag checks to ensure that the EyeDropper
is not created when it is not supported.

[1]: https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/ui/views/eye_dropper/eye_dropper.cc

Bug: 1247514
Change-Id: I2854d19d081367b3c3c535553bbdfe5fb0a86b3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3149734
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#919459}
NOKEYCHECK=True
GitOrigin-RevId: 7329e1bbadff941d595566705b34af6f9b68941f