Fixing potential OOM with OffscreenCanvas
Recently to fix an issue with clearing some OffscreenCanvases with
pending activity, we added some code to protect them from being Garbage
Collected if they still had things to draw on screen.

The issue was if these canvas were not displayed on the screen in the
first place, so we add now more code to handle this case.

This CL also addresses a pending TODO regarding using Host instead of
offscreenCanvasForBinding.

Bug: 1247157
Change-Id: Ia282bb7ed6a4ed00fb37fc6799de7add7c0dfdfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3149275
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919466}
NOKEYCHECK=True
GitOrigin-RevId: 2d25e98fd9882bbe2eaff98cca8573ad48e057bc