Track by Node instead of DOMNodeId in TextPaintTimingDetector
Currently, the map containing the TextRecords used in LCP and in Element
Timing use DOMNodeId. This CL changes this behavior to use a
Node weak pointer. Besides removing the need to compute node ids for
nodes that do not already have them, this is also a first step towards
a fix for recomputing the LCP size once a text node paints with its
final font. Since we'd like to update the size when this happens, we
need to easily determine if a given LayoutObject is the LCP candidate,
and using the DOMNodeId in the map prevents us from doing this
efficiently.

Bug: 1213220
Change-Id: Id432ef20aeaf45149813ddbc8a842a6ea1d6582a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3145538
Commit-Queue: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919499}
NOKEYCHECK=True
GitOrigin-RevId: 63333c554f4e0d7a53df9084832f567230914c3b