[@container] Delayed animation updates
This CL implements a new approach [1][2] for dealing with animated
container queries. The approach can be summarized as:

  1) Delaying application of the CSSAnimationUpdate until after
     layout, and
  2) Remembering the "original" old-styles for relevant elements.

Both of these things were also required for the previous solution
(currently behind the runtime flag CSSIsolatedAnimationUpdates),
so they are already mostly in place. Apart from these two things,
however, the new solution actually has more in common with how
things already work on stable than with CSSIsolatedAnimationUpdates,
so this this CL effectively reverts that code.

The new approach in this CL works like this:

 - During StyleResolver::ResolveStyle, elements that have a non-
   empty animation update gets added to a set on DocumentAnimations.
 - A stack-allocated object CSSAnimationUpdateScope is placed on the
   sites where we do style and layout, and it ensures that any
   collected pending updates are applied before we exit that scope.
 - Old-styles are temporarily added for each element we resolve
   style for, with a lifetime that's also tied to
   CSSAnimationUpdateScope.

Regarding InApplyAnimationUpdateScope: it turns out that applying
an animation update can cause calls to Element::
SetNeedsAnimationStyleRecalc. On main, these calls are ignored,
because we always apply the animation update during style recalc,
and we early-out from SetNeedsAnimationStyleRecalc if we're currently
in style recalc. With the changes in this CL, we apply the same
update *after* style recalc, so we need another way to early-out
from SetNeedsAnimationStyleRecalc, and this is what the
InApplyAnimationUpdateScope is for.

Note: It probably makes sense to store the elements-with-pending-
      updates set and the old-styles on the CSSAnimationUpdateScope
      itself, but I'll try that in a separate CL.
Note: We still store old-styles for way too many elements. An
      improvement for that is coming soon.

[1] https://docs.google.com/document/d/1Ef9V-Ra_6I3R9DY0o10fZ7HTriPm9oHYw1xUN0bRcHw
[2] https://github.com/w3c/csswg-drafts/issues/6398#issuecomment-904754453

Change-Id: I735b9113c12ef57696b82470b2256c639884ef48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3129367
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919696}
NOKEYCHECK=True
GitOrigin-RevId: 0548dd9d10add21f75e4685e91389ff72dc99a5d