Enable Origin Trial for Access Handles
This CL replaces the content flag that enables Access Handles for a
runtime one, and then adds an Origin Trial for the feature. Changing the
flag types greatly simplifies OT integration, and the original reason to
use a content flag is no longer relevant (see
https://chromium-review.googlesource.com/c/chromium/src/+/2949473/24#message-05c8e0d4de50f4189c6a76b17ff1bdd9f3b3e6f5,
a content flag was suggested to add extra security to the new mojo
interfaces. We have undergone a security review and during an OT the
origin can choose to enable the feature anyway). This CL also adds
measurements to all access-handle-related methods.

The launch bug for this origin trial is:
https://bugs.chromium.org/p/chromium/issues/detail?id=1232436

Bug: 1218431
Change-Id: Iddf6da9e54ee55dde1a9ff207ba45d3c97b93a81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3141881
Commit-Queue: Emanuel Krivoy <krivoy@google.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Austin Sullivan <asully@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919744}
NOKEYCHECK=True
GitOrigin-RevId: 6aeff72af52575cd9d3edb1d2a6ba4518aba8825