Add FileSystemHandle::move() and FileSystemHandle::rename() methods
Currently, it is not possible to move or rename a file or directory
without creating a new file/directory, copying over data (recursively,
in the case of a directory), and removing the original.

This CL allows for the atomic moving of a file or directory on the
local file system without needing to duplicate data.

Moves to non-local file systems will are not guaranteed to be atomic
and will involve duplicating data.

PR: https://github.com/WICG/file-system-access/pull/317

Bug: 1140805
Change-Id: I774ed1d9616249b6ecc80783db48a7bfee915aab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2984739
Commit-Queue: Austin Sullivan <asully@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919810}
NOKEYCHECK=True
GitOrigin-RevId: 89be8a84bf5c962fa363bd093f4f125ac923838c