Avoid suboptimal breaks inside a column row when appropriate.
Set a minimum appeal for breaks inside a nested multicol container, if
the row is past the block-start of the outer fragmentainer. If there's
content that breaks suboptimally (with a lower appeal than any previous
breaks), push such content ahead of us, so that it ends up in the next
outer fragmentainer. There may be more space there, so that the
suboptimal break may be avoided.

Propagate the lowest appeal among all column breaks to the multicol
fragment builder, so that an ancestor can decide to break before the
entire multicol container, if that would be better. This currently only
works if the multicol container itself breaks, since break appeal is
stored in the break tokens. A follow-up CL will fix this (by storing it
in NGLayoutResult instead), so that we'll also detect suboptimal breaks
inside a nested multicol container that didn't break in the outer
fragmentation context.

This doesn't fix any existing tests by itself, but will contribute to
it. A follow-up CL will fix some remaining problems, so that quite a few
existing tests will finally pass.

Bug: 829028
Change-Id: I4a09e2163aa194780be986f358283d4e8da9a115
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3149995
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919814}
NOKEYCHECK=True
GitOrigin-RevId: b12ea5125ee979cf9cdd35b5b652d68b36bf45cd