[Blink] Introduce a public constant for the main world ID
The main execution world for a document has a reserved ID of `0`. This
was only defined in blink, but was also used (at least) in extensions
code. Introduce a new constant, blink::kMainDOMWorldId, for embedders
to use, and static_assert() that it's the same value as
DOMWrapperWorld::kMainWorldId.

Bug: None
Change-Id: Ie02c11997dac5b16972c0a2c3240fbfac09e8253
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3150537
Commit-Queue: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919931}
NOKEYCHECK=True
GitOrigin-RevId: f1f62c835b736af66e3886b1a02f1ac484adc30c