webhid: Close service connection when context is lost
This fixes an issue where the execution context could be destroyed
before the mojo connection to the HID service is disconnected, causing
blink::HID to be torn down with pending Promise callbacks.

This CL makes blink::HID an ExecutionContextLifecycleObserver so it
can respond to ContextDestroyed and resolve the pending callbacks.

Bug: 1243284
Change-Id: Ic1860ffa39d7e1fc7197b14c8cb0bbe4570d1115
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3140735
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919947}
NOKEYCHECK=True
GitOrigin-RevId: 0c75e88e66cd8232b1e5d0624d9039c5b90aaebb