Assert that we have enough NGPhysicalBoxFragment objects.
For any LayoutBox that is to participate in fragment traversal, we
require that the number of NGPhysicalBoxFragment and FragmentData
objects be identical. This assumption is made in
PaintLayer::CollectFragments(), and we currently have a crash because of
it (calling LayoutBox::GetPhysicalFragment() with an invalid index).

We'll now fail a DCHECK after pre-paint if this isn't the case, in the
hope that this will now fail more often (for objects that don't
establish a PaintLayer, and also for objects that don't get painted,
because they're outside the viewport), which in turn might lead to a
reproducible test case from somewhere (e.g. a fuzzer).

Bug: 1244130
Change-Id: I8a8567835fe45cada793fd463bae703b34bd2c1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3138497
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#920312}
NOKEYCHECK=True
GitOrigin-RevId: 7565fba2d761a7403930c7fea939c2b1cac77c30