Relanding: Integrated Progress based animations with CSS Animations
This is a relanding of:

https://chromium-review.googlesource.com/c/chromium/src/+/3130703

The only differences between this CL and the original is removal of a
time-range related test case and the change in:
third_party/blink/web_tests/external/wpt/scroll-animations/current-time-root-scroller.html

The fix was to add a single frame in between creating the timelines and
trying to access them. This method is also used in many other existing
tests.

Original description:

Timeline duration can now be stored on an AnimationEffect.
This was needed since for CSS animations, InertEffects with
no owning animation are used until later in the pipeline.
Normally the owning animation of an effect is used to gain
access to the corresponding timeline which provides the timeline
duration needed for timing normalization math, but in some
cases the duration was needed before an effect had an owner
assigned. As such it is now possible to create an InertEffect
and pass in a timeline duration if needed.

Added logic for when CSS animations swap from progress based
timeline to a time based timeline. This new logic preserves
the current progress of the animation even when current times
do not match up.

I also removed time range from some CSS tests since they are
now supported by progress based animations.

Bug: 1216655, 1140602, 1243947
Change-Id: Ibec9437c1027702b888e8bab65f890687054e3be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3142895
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Jordan Taylor <jortaylo@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#920356}
NOKEYCHECK=True
GitOrigin-RevId: c1ffb982fb8d25e4541bb42254f64dae98004398