[GridNG] Fixing NGDisableSideEffectsScope in ComputeMinMaxSizes
In https://chromium-review.googlesource.com/c/chromium/src/+/2991438,
an NGDisableSideEffectsScope element was added, intended only for the
ComputeMinMaxPhase. However, it was gated on
"sizing_constraint != SizingConstraint::kLayout", which is not correct,
since the row computation in ComputeMinMaxSizes is correctly using a
sizing_constraint of SizingConstraint::kLayout.

The fix here is to plumb a boolean down to indicate whether we're in
the ComputeMinMaxSizes stage.

Bug: 1248162
Change-Id: I83c9268311f1e1b343d32e07f1cc1195ceab9447
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3154996
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#920442}
NOKEYCHECK=True
GitOrigin-RevId: 71242913272fc77fa14ba11380cab96437259f99