Enable OneCopyCanvasCapture with no GrContext
OneCopyCanvasCapture was broken when used with OOP-R canvas due to
the absence of a renderer-side GrContext. This has been implemented in
crrev.com/914885, so it can be re-enabled.

Rename the feature to TwoCopyCanvasCapture. Two reasons. First, one of
the copies is still alive and well. Second is that the Finch trial that
is enabling OOP-R canvas is disabling OneCopyCanvasCapture, because
this GrContext null-check was missing.

Also update SharedImageRepresentationSkiaImpl::BeginWriteAccess to
avoid spamming the console with warnings about converting trying to
create a YUV SkColorSpaces.

Bug: 1206168
Change-Id: Ide3f973b3fe48c23251135f3374c3be292e53248
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3150309
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Kenneth Russell <kbr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#920480}
NOKEYCHECK=True
GitOrigin-RevId: dd4c86e36cf98eedde96f27735c147d4003a84f1