[editing] Avoid DCHECK failure in ApplyStyleCommand::RemoveInlineStyle
This function was not considering that the start and end positions of
the range could be equal. In that case, when removing the anchor, it
would update the start position |s| to be at the beginning of the next
node, and the end position |e| to be at the end of the previous one.

That could fail this assert:

  DCHECK(s.IsAfterAnchor() ||
         !OffsetIsBeforeLastNodeOffset(s.OffsetInContainerNode(),
                                       s.ComputeContainerNode()));

and later try to create |EphemeralRange(s, e)| with |e| preceding |s|.

This patch changes it so that, if the positions are equal, both are set
to be between the previous and the following node of the removed one.

Bug: 1246190

TEST=ApplyStyleCommandTest.RemoveEmptyItalic

Change-Id: If1ddda73138b17c6fda96ff488908294f36ef037
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3156002
Commit-Queue: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#920686}
NOKEYCHECK=True
GitOrigin-RevId: 8668ce9ab37aaa078ef9efef4a6a97045dca636b