[editing] Avoid ReplaceSelectionCommand crashes with mixed editability
Executing document.execCommand('InsertImage') with this HTML:

  <body contenteditable="true">
    <div contenteditable="false">
      <span contenteditable="true">a|b</span>
    </div>
  </body>

runs ReplaceSelectionCommand::InsertParagraphSeparatorIfNeeds().
This would typically split the <span> and insert the <img> as a child of
the <div>. However, the <div> is not editable, so a new <div> was added
at the end of the <body>, since it's an editable ancestor.

A DCHECK failed when collapsing the selection to the VisiblePosition in
the new <div> because the canonical position is null when the height of
a block is 0 (see IsVisuallyEquivalentCandidateAlgorithm).

Alternatively, we can modify the testcase a bit to trigger a null deref.
MergeEndIfNeeded() uses PositionAtEndOfInsertedContent(), which may be
a null VisiblePosition if the original Position couldn't be normalized.
Then it could call InsertNodeBefore() with a null node which would be
dereferenced.

This patch avoids this kind of problems by making that code not skip
over non-editable ancestors in EnclosingBlock(). To do so, changes
EnclosingNodeOfTypeAlgorithm() to use RootEditableElementOf() instead
of HighestEditableRoot() when rule == kCannotCrossEditingBoundary.
The difference is that RootEditableElementOf() doesn't cross editing
boundaries.

Bug: 1246674

TEST=ReplaceSelectionCommandTest.InsertImageInNonEditableBlock1
TEST=ReplaceSelectionCommandTest.InsertImageInNonEditableBlock2
TEST=external/wpt/editing/run/insertparagraph.html?6001-last

Change-Id: Icf93a26c6f1581fc1710a45128fe36d8860e5701
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3156666
Commit-Queue: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#920693}
NOKEYCHECK=True
GitOrigin-RevId: f4e774a6bf195fc3317c5fc83ab9d02bf585e77b