[inert] Ignore 'inert' attribute on non-HTML elements
'inert' is an HTML attribute so it should have no effect when directly
used on other kinds of elements like SVG or MathML.

But of course, if 'inert' is set on an HTML ancestor element, then SVG
and MathML will still be inert, can't escape the inert subtree.

This change aligns Blink with Gecko.

Bug: 692360

TEST=external/wpt/inert/inert-on-non-html.tentative.html

Change-Id: If19ef9cd5afa06da1e42d1fd2956a5dbc4c5a1f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3157603
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/main@{#920741}
NOKEYCHECK=True
GitOrigin-RevId: 3fee8d24b111b9bd6e96d88773f67e336addd40f