[SelectMenu] Allow the UA provided shadow DOM to be replaced via attachShadow.
This CL enables authors to replace the UA provided shadow DOM
for <selectmenu> by calling attachShadow().

For now this change follows the behavior of declarative shadow DOM
where the initial shadow DOM is emptied out and attachShadow returns
a reference to it. There is an open-ui issue[1] to further discuss this
behavior.

The initial properties of the UA shadow root are updated with
the ones provided by attachShadow.

[1]: https://github.com/openui/open-ui/issues/399

Bug: 1121840
Change-Id: I91758022b211fd4c511c0f86997de347c03e6b83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3151015
Reviewed-by: Dan Clark <daniec@microsoft.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Ionel Popescu <iopopesc@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#920836}
NOKEYCHECK=True
GitOrigin-RevId: f0c3efe4e884df8226b3a3a2dff98dec86597f6d