[editing] Handle null positions in PositionIterator
PositionIterator (or PositionIteratorInFlatTree) can be constructed with
a Position (or PositionInFlatTree). The code was not checking whether
the position is null, triggering null derefs.

This patch delegates to the PositionIteratorAlgorithm() constructor to
avoid null derefs, and also fixes DeprecatedComputePosition() and
ComputePosition() methods to avoid more null derefs.

Bug: 1248744

TEST=PositionIteratorTest.nullPosition

Change-Id: I12760e57aba2ada7cfa86eec2b006fa8151bb0f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3157612
Commit-Queue: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#921072}
NOKEYCHECK=True
GitOrigin-RevId: 6113a4732f9c9d0b34bb3fe95950fe4c944f27a1