Make SelectionDisplayItemClient GarbageCollected
DisplayItemClient is inherited by LayoutObject, which was moved to
Oilpan.
It is necessary to make it GarbageCollected because it’s dangerous to
save a raw pointer to an on-heap object.
Every child class of DisplayItemClient should be GarbageCollected to
make it possible.

Bug: 1248357
Change-Id: I499bf90e5d6207f284783c32fe27dfa7bc8045aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3156309
Reviewed-by: Yuki Yamada <yukiy@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Naho Tomiki <nahotomiki@google.com>
Cr-Commit-Position: refs/heads/main@{#921081}
NOKEYCHECK=True
GitOrigin-RevId: 4bba7ba7aeeb5e2d21b15798bc2834562166be3f