Make SameSizeAsInlineBox GarbageCollected
DisplayItemClient is inherited by LayoutObject, which was moved to
Oilpan.
It is necessary to make it GarbageCollected because it’s dangerous to
save a raw pointer to an on-heap object.
Every child class of DisplayItemClient should be GarbageCollected to
make it possible.

Bug: 1248357
Change-Id: I90869b1333fec5ae47bb907bae4aaedb8f65161d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3157824
Reviewed-by: Yuki Yamada <yukiy@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Naho Tomiki <nahotomiki@google.com>
Cr-Commit-Position: refs/heads/main@{#921082}
NOKEYCHECK=True
GitOrigin-RevId: d7d55a6a0f1cace81e3b15336057cd8d1201a779