[@scroll-timeline] Effectively pause animations in flaky WPT
The last test in at-scroll-timeline-dynamic.tentative.html expects
*zero* time to have passed between the first and second calls to
'await assert_width(...)', when in reality several frames worth of
time has passed. This causes flakiness on bots.

We don't actually care about exactly where we're at on the document
timeline in this test file, only whether or not we are currently on
that timeline. Hence this CL uses the steps() function to effectively
"stop" document-timeline-attached animations for a full minute.

Fixed: 1248912
Change-Id: I53b0c4319daf306c0e664454b1b48208c384c657
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3158007
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#921117}
NOKEYCHECK=True
GitOrigin-RevId: a093e8e1759284a5a3cfb8337be491c435a3c6a5