Making GetImageWithAlphaDisposition early return if image is Opaque
GetImageWithAlphaDisposition was recently starting to be used with
Accelerated Canvases. This method was not handling correctly when the
canvas is marked as Opaque.

This CL adds an early return for GetImageWithAlphaDisposition when the
image is opaque.

This CL also optimizes GetImageWithAlphaDisposition to not having
repeated calls to PaintImageForCurrentFrame.

Bug: 1248818
Change-Id: Ie2606d4a6f6e3c41e6b1d964511e535ae795f3b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3158902
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#921219}
NOKEYCHECK=True
GitOrigin-RevId: 04b79157b668d77fc54c92fb1b44d687ee844f22