[Web Payment] Add UserCounters for completing PaymentRequest.
Before this patch, it was possible to monitor PaymentRequest completion
rates only in the browser process, which may miss some issues in code
that cause an abort before the measurement code is reached.

This patch adds UseCounters for PaymentReqhest.show() resolving and
PaymentRequest.complete() being called in Blink.

After this patch, even if the completion rate measurement code in the
browser process is not reached due to some issue, the Blink UseCounters
can be used as a backup.

Bug: 1203257
Change-Id: I21c735277c54ddbd92ceab3c4325de4c6e865439
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3160859
Reviewed-by: Liquan (Max) Gu <maxlg@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#921407}
NOKEYCHECK=True
GitOrigin-RevId: 37bf64e24c2aedb2d2cb8b58125a3d9b29babe41