Do not throttle BeginFrame during Picture in Picture
When Picture in Picture is rendering content that's being drawn by the
page, rather than directly from a video source, blink may emit
CompositorFrames while it's backgrounded as part of creating that
content.  If the page is backgrounded, this can cause viz to throttle
OnBeginFrame (== requestAnimationFrame and requestVideoFrameCallback),
and the picture in picture window updates at the throttled ~1fps rate.

This CL works around it by allowing blink to request that it is not
throttled while it is driving Picture in Picture.  It may be better
to do something else, such as preventing those frames from being
sent to viz, but this is much easier to reason about.

As an aside, VideoFrameSubmitter's frame sink can still be throttled,
since it doesn't mark its CompositorFrames with the new "do not
throttle" request.

For this bug, that's okay, but it means that post-processing a src=
or MSE video by the page might still be throttled.  MediaStream
isn't affected, due to implementation differences.  No attempt was
made to fix this, since un-throttling VideoFrameSubmitter doesn't
help the issue observed here, and would add needless traffic to
viz in the cases we do care about.

Change-Id: I76fd802f7e8b759ca04b3e48117da7594d6e6f4d
Bug: 1232173
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3138707
Reviewed-by: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#921442}
NOKEYCHECK=True
GitOrigin-RevId: 58ad90a239fd4cba8bd37e01a2c40055b8609727