[LayoutNG] Fieldset container separation
There should never be a class C breakpoint in front of fieldset
content. As a result, always pass in false for
|has_container_separation| when evaluating a break before fieldset
content.

As a result of this update, if the legend overflows the fragmentainer,
we may now end up pushing the entire fieldset to the next fragmentainer,
as expected. (See the newly added WPT test: fieldset-006.html).

Several unit tests were updated to reflect this change in behavior.

Bug: 1246175
Change-Id: I1e6e0393119fe01c6a293fb0d09f4a779684e9d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3139496
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#921481}
NOKEYCHECK=True
GitOrigin-RevId: 435aa69bbd6a76bb72f46ea025dc18e1085ea499