[Autofill] In preview mode, only highlight changed fields
Currently, Autofill highlights all fields that are either already
autofilled or a filling is currently previewed.

With this CL the behavior is changed:

a) If Autofill is NOT in preview mode, all autofilled fields are
highlighted which is the current behavior.

b) If Autofill is in preview mode, only the fields that would change
by accepting the filling are highlighted.

To achieve b), fields that are already in the autofilled state must not
be highlighted in preview mode. For this a binary state for form control
elements is introduced that indicates if a
WebAutofillState::kAutofilled field should be highlighted.
This field is evaluated for selecting the css pseudo class and for
painting the fields.

Change-Id: I0d633bf4c480b061d6c8bcb45405514306ab7fd9
Bug: 1248585, 1247978
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3151850
Commit-Queue: Matthias Körber <koerber@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#921756}
NOKEYCHECK=True
GitOrigin-RevId: 44082a4934c96e1ea950ecf88be1cefbae265a03