Unify canvas 2d context auto-flushing logic
Before this CL we had auto-flushing logic in MemoryManagedPaintCanvas
that triggered a flush when the RAM pinned by source images reaches a
threshold, and we had separate logic in CanvasResourceProvider that
triggered a flush when the Op count or the RAM consumed by serialized
ops crossed a threshold.

This CL brings together these these two mechanisms, to centralize them
in CanvasResourceProvider.  The callback that was used by
MemoryManagedPaintCanvas to trigger flushes was replaced with a callback
that reports memory consumption. The callback mechanism that use to be
based on base::Callback was overkill and was replaced with a leaner
client virtual interface.

Moving the logic to CanvasResourceProvider was essential for makin the
flushing logic safe with respect to the new beginLayer/endLayer API.
This is because the 2d context state is not accessible from
MemoryManagedPaintCanvas

Also, several unit tests were added to verify that the unified
auto-flushing logic plays nice with overdraw optimizations and
beginLayer/endLayer

Bug: 1249189
Change-Id: Ib9b7c4bd4943990dbcaaaf25ddb0c5047f66aa33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3163396
Commit-Queue: Justin Novosad <junov@chromium.org>
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#921895}
NOKEYCHECK=True
GitOrigin-RevId: 2762cb09725320da7cfa57020933d504d2a25532