Make PaintRecordBuilder GarbageCollected
DisplayItemClient is inherited by LayoutObject, which was moved to
Oilpan.
It is necessary to make it GarbageCollected because it’s dangerous to
save a raw pointer to an on-heap object.
Every child class of DisplayItemClient should be GarbageCollected to
make it possible.

Bug: 1248357
Change-Id: I6d40c4b84d34b16631ea8f87026c99b66d2ed1f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3159941
Reviewed-by: Yuki Yamada <yukiy@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Naho Tomiki <nahotomiki@google.com>
Cr-Commit-Position: refs/heads/main@{#921967}
NOKEYCHECK=True
GitOrigin-RevId: 16939a7e6f83c65a067b6105913b71af558eeabb