Make FakeDisplayItemClient GarbageCollected
DisplayItemClient is inherited by LayoutObject, which was moved to
Oilpan.
It is necessary to make it GarbageCollected because it’s dangerous to
save a raw pointer to an on-heap object.
Every child class of DisplayItemClient should be GarbageCollected to
make it possible.

Bug: 1248357
Change-Id: Id8b7e40793c58315a71d3f4aec77ce81a3c56ca9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3159383
Commit-Queue: Naho Tomiki <nahotomiki@google.com>
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Yuki Yamada <yukiy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#921983}
NOKEYCHECK=True
GitOrigin-RevId: c2e183898bb65ea7bf69177349342ba06f02bc81