[wasm] Disable failing WPT tests
The last CL disabled only the main-thread version of the tests but not
the worker version.

Also, an update of the wpt tests re-enabled some wpt tests that are
blocking the launch of WebAssembly Reference Types. This CL disables
these tests again.

R=mlippautz@chromium.org

Bug: v8:12227
Change-Id: I66bd09f819b6dc22c3ca0c1f52a323903ec31e62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3162980
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#922024}
NOKEYCHECK=True
GitOrigin-RevId: 4a0c76f58644ba236b38bb5a6ca7c401f6abd827